fix: move migration
